Implemented all features. Refactored packages. Needs commenting and beautification
