+++ app/src/main/java/com/example/compassalpha/SettingsActivity.java
@@ -5,6 +5,9 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.widget.Switch;
 
+import com.example.compassalpha.nav.NavHandler;
+import com.example.compassalpha.settings.Settings;
+
 public class SettingsActivity extends AppCompatActivity {
     private static final String tag = "SettingsActivity";
