+++ app/src/main/java/com/example/compassalpha/compass/Compass.java
@@ -1,4 +1,4 @@
-package com.example.compassalpha;
+package com.example.compassalpha.compass;
 
 import android.content.Context;
 import android.hardware.Sensor;
