+++ app/src/main/java/com/example/compassalpha/compass/CompassLogic.java
@@ -1,4 +1,4 @@
-package com.example.compassalpha;
+package com.example.compassalpha.compass;
 
 import android.hardware.SensorManager;
