+++ app/src/main/res/layout/activity_main_content.xml
@@ -29,7 +29,7 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/azimuth_direction" />
 
-    <com.example.compassalpha.CompassView
+    <com.example.compassalpha.compass.CompassView
         android:id="@+id/compass_view"
         android:layout_width="380dp"
         android:layout_height="398dp"
